Little code tweaks, nothing signficant. Ran the whole app through a bug sniffer and fixed everything it found. Mostly got rid of unneccessary casts and unused calls.  Made brain and brainspec abstract.

+++ src/sides/SideReader.java
@@ -173,14 +173,14 @@ public class SideReader {
 		Color color = Color.white;
 		try {
 			if (token.length() == 3) {
-				r = (float) Integer.parseInt(token.substring(0, 1), 16) / 16.0f;
-				g = (float) Integer.parseInt(token.substring(1, 2), 16) / 16.0f;
-				b = (float) Integer.parseInt(token.substring(2, 3), 16) / 16.0f;
+				r = Integer.parseInt(token.substring(0, 1), 16) / 16.0f;
+				g = Integer.parseInt(token.substring(1, 2), 16) / 16.0f;
+				b = Integer.parseInt(token.substring(2, 3), 16) / 16.0f;
 				color = new Color(r, g, b);
 			} else if (token.length() == 6) {
-				r = (float) Integer.parseInt(token.substring(0, 2), 16) / 255.0f;
-				g = (float) Integer.parseInt(token.substring(2, 4), 16) / 255.0f;
-				b = (float) Integer.parseInt(token.substring(4, 6), 16) / 255.0f;
+				r = Integer.parseInt(token.substring(0, 2), 16) / 255.0f;
+				g = Integer.parseInt(token.substring(2, 4), 16) / 255.0f;
+				b = Integer.parseInt(token.substring(4, 6), 16) / 255.0f;
 				color = new Color(r, g, b);
 			}
 			return color;
